move SnapshotWithMutex to an individual file
